Add post: Exploring databases in Vim with vim-dadbod
Introduced a new blog post that delves into the `tpope/vim-dadbod` Vim
plugin, which allows users to connect to databases and execute SQL
queries directly within the editor, eliminating the need for a GUI or
context switching. The post covers basic usage, including connecting to
a local SQLite database, browsing tables, and executing queries with
ease. Additionally, it highlights the integration with Vim plugins like
`vim-dadbod-ui` and `vim-dadbod-completion`, making it a must-have for
users heavily reliant on a terminal-based workflow. [ci skip]